requirements.txt
spotipy
youtube-dl
google-api-python-client
yt-dlp
flask
python-dotenv